feat(#213): Activity를 Fragment로 변경 및 navigation 추가
