+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/profile/ProfileViewModelFactory.kt
@@ -1,0 +1,16 @@
+package boostcamp.and07.mindsync.ui.profile
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import boostcamp.and07.mindsync.data.repository.profile.ProfileRepository
+
+class ProfileViewModelFactory(private val profileRepository: ProfileRepository) :
+    ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(ProfileViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return ProfileViewModel(profileRepository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
