+++ AOS/app/src/main/res/navigation/nav_profile.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_profile"
+    app:startDestination="@id/profileFragment">
+
+    <dialog
+        android:id="@+id/editNickNameDialog"
+        android:name="boostcamp.and07.mindsync.ui.dialog.EditNickNameDialog"
+        android:label="EditNickNameDialog"
+        tools:layout="@layout/dialog_edit_profile"/>
+    <fragment
+        android:id="@+id/profileFragment"
+        android:name="boostcamp.and07.mindsync.ui.profile.ProfileFragment"
+        android:label="ProfileFragment"
+        tools:layout="@layout/fragment_profile">
+        <action
+            android:id="@+id/action_profileFragment_to_editNickNameDialog"
+            app:destination="@id/editNickNameDialog" />
+    </fragment>
+</navigation>
